Squashing migration files for v3
